pyproject.toml
[project]
name = "connectdb"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "flask>=3.1.0",
    "flask-sqlalchemy>=3.1.1",
    "psycopg2>=2.9.10",
    "python-dotenv>=1.0.1",
    "sqlalchemy>=2.0.39",
]
